Locally functioning map
Made it so markers stay after createAlert, if you're operating on a single machine. Basically if you do a report, you will see it as long as you stay on one of the websites in the same machine. But someone on a different machine won't see your reports/markers, and you won't see their reports/markers